Add: memberId 자료형 Long으로 변경

+++ src/main/java/couponToy/CouponToyProject/Member/model/Member.java
@@ -13,7 +13,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 public class Member {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer memberId;
+    private Long memberId;
 
     @Column(nullable = false)
     private String email;
